refactor: split queries.ts into domain-focused modules
Split the 2078-line services/queries.ts into a queries/ directory with
focused modules: types, metrics, daily-summary, period-summary, tags,
activities, productivity, and locations. Barrel index.ts re-exports all
original public symbols so consumers are unaffected.
