+++ apps/backend/src/routes/metrics-router.ts
@@ -62,7 +62,7 @@ import {
   queryMetrics,
   queryMetricsBucketed,
   type SyncProvider,
-} from '../services/queries.ts'
+} from '../services/queries/index.ts'
 import { getLatestMetric } from '../services/reports.ts'
 import { type TypedRouter, typedRouter } from '../typed-router.ts'
 import { validateBody, validateQuery } from '../validation.ts'
